# Performance Optimization: Achieve 95%+ Lighthouse Scores