--- a/package.json
+++ b/package.json
@@ -22,6 +22,7 @@
     "sharp": "^0.33.5"
   },
   "devDependencies": {
+    "@next/bundle-analyzer": "^15.5.2",
     "@types/google.maps": "^3.58.1",
     "@types/lodash": "^4.17.13",
     "@types/node": "^20.11.17",
@@ -33,6 +34,7 @@
     "cssnano": "^7.1.1",
     "eslint": "^8.57.1",
     "eslint-config-next": "^14.1.0",
+    "lighthouse": "^12.8.2",
     "postcss": "^8.4.35",
     "tailwindcss": "^3.4.1",
     "typescript": "^5.8.3"