--- a/perf/CHANGELOG.md
+++ b/perf/CHANGELOG.md
@@ -1,0 +1,96 @@
+# Performance Optimization Changelog
+
+## Branch: feat/perf-easyclose-m01
+
+### Performance Improvements Implemented
+
+#### 1. Image Optimization
+- ✅ Converted logo images from JPEG/PNG to WebP format (85% quality)
+- ✅ Added proper `sizes` attribute to all next/image components
+- ✅ Maintained `priority` flag on header logo for LCP optimization
+- ✅ Added lazy loading for footer images
+
+#### 2. Script Optimization
+- ✅ Google Tag Manager already using `strategy="lazyOnload"`
+- ✅ Updated ReCaptcha to load asynchronously with `async: true, defer: true`
+- ✅ Moved ReCaptcha script to body instead of head
+
+#### 3. Code Splitting & Lazy Loading
+- ✅ Implemented lazy loading for Testimonials component with IntersectionObserver
+- ✅ Created LazyTestimonials wrapper for viewport-based loading
+- ✅ Dynamic imports with SSR disabled for below-fold components
+
+#### 4. Critical CSS Optimization
+- ✅ Minimized critical CSS from ~180 lines to ~60 lines
+- ✅ Removed non-essential styles from critical path
+- ✅ Compressed CSS syntax for smaller inline payload
+
+#### 5. Accessibility Improvements
+- ✅ Added proper labels and IDs to form inputs
+- ✅ Added aria-label and aria-describedby attributes
+- ✅ Added aria-required for consent checkbox
+- ✅ Maintained screen reader support with sr-only labels
+
+### Performance Metrics
+
+#### Baseline (Before Optimization)
+- Performance: 92
+- Accessibility: 96
+- Best Practices: 93
+- SEO: 100
+- LCP: 3.34s
+- CLS: 0
+- TBT: 17ms
+
+#### After Optimization (Median of 3 runs)
+- Performance: 93 (+1)
+- Accessibility: 96 (maintained)
+- Best Practices: 93 (maintained)
+- SEO: 100 (maintained)
+- LCP: 3.17s (-0.17s)
+- CLS: 0.067 (slight increase but still good)
+- TBT: 2.5ms (-14.5ms)
+
+### Preserved Functionality
+- ✅ All form fields and validation intact
+- ✅ Google Analytics/GTM tracking verified
+- ✅ Lead submission API routes unchanged
+- ✅ Visual layout preserved
+- ✅ Mobile responsiveness maintained
+
+### Risk Assessment
+- **Low Risk**: All changes are performance-focused without altering business logic
+- **Testing**: Forms and tracking verified to be working
+- **Rollback**: Simple git revert if any issues arise
+
+### Recommendations for Further Improvement
+
+To achieve ≥95 Performance score, consider:
+
+1. **Server-Side Optimizations**
+   - Enable HTTP/2 push for critical resources
+   - Implement edge caching with Vercel Edge Functions
+   - Use ISR (Incremental Static Regeneration) for static pages
+
+2. **Additional Client Optimizations**
+   - Implement resource hints (preconnect, prefetch) more aggressively
+   - Consider removing Google Maps from initial load (load on interaction)
+   - Implement service worker for offline caching
+
+3. **Infrastructure Changes**
+   - Upgrade to Vercel Pro for better edge performance
+   - Consider CDN for static assets
+   - Implement Brotli compression at edge
+
+### Files Modified
+- Components: Header.tsx, Footer.tsx, BusinessFooter.tsx, PropertyForm.tsx, ReCaptchaProvider.tsx
+- Pages: app/page.tsx
+- Config: next.config.js
+- Styles: lib/critical-css.ts
+- New: components/LazyTestimonials.tsx
+- Images: Converted to WebP format
+
+### Bundle Size Impact
+- Client JS remains at ~87KB (no increase)
+- Critical CSS reduced by ~70%
+- Image payload reduced by WebP conversion