… perf/bundle-before.html
… perf/lighthouse-after-1.json
… perf/lighthouse-after-2.json
… perf/lighthouse-after-3.json
… perf/lighthouse-before-1.json
… perf/lighthouse-before-2.json
… perf/lighthouse-before-3.json
… perf/lighthouse-final-1.json
… perf/lighthouse-final-2.json
… perf/lighthouse-final-3.json