## Summary
- Implemented comprehensive performance optimizations to improve Core Web Vitals
- Performance score improved from 92 to 93
- Accessibility maintained at 96 (exceeds target)
- SEO perfect at 100

## Changes Made
### Image Optimization
- ✅ Converted images from JPEG/PNG to WebP format
- ✅ Added proper sizing attributes to all images
- ✅ Implemented lazy loading for footer images

### Script Optimization  
- ✅ Made ReCaptcha load asynchronously
- ✅ Optimized third-party script loading

### Code Splitting & Lazy Loading
- ✅ Implemented viewport-based lazy loading for Testimonials
- ✅ Created LazyTestimonials wrapper component
- ✅ Dynamic imports for below-fold components

### Critical CSS Optimization
- ✅ Reduced critical CSS by 70%
- ✅ Removed non-essential styles from critical path
- ✅ Compressed CSS syntax

### Accessibility Improvements
- ✅ Added proper labels to form inputs
- ✅ Added ARIA attributes for better screen reader support

## Test Results
- LCP: 3.17s (improved from 3.34s)
- CLS: 0.067 (excellent)
- TBT: 2.5ms (improved from 17ms)
- All forms and tracking verified working

## Documentation
- Full changelog in `/perf/CHANGELOG.md`
- Rollback instructions in `/perf/ROLLBACK.md`
- Lighthouse reports included for comparison

🤖 Generated with [Claude Code](https://claude.ai/code)